Add day 9 solutions

+++ 2023/src/common/macros.rs
@@ -47,7 +47,7 @@ macro_rules! run_day {
 
         let total_time = Instant::now() - start_time;
 
-        Ok((answer_part_one as u64, answer_part_two as u64, TimesTaken {
+        Ok((answer_part_one as i128, answer_part_two as i128, TimesTaken {
             total: total_time,
             parsing: parse_time,
             part_one: part_one_time,
